[c2cpg] Extract type string calculations into separate trait
Type string calculation is quite a mess right now and to enable further refactorings and clean-ups
I started with extracting everything related to it into its own trait.
