+++ joern-cli/frontends/c2cpg/src/main/scala/io/joern/c2cpg/astcreation/AstCreator.scala
@@ -2,14 +2,18 @@ package io.joern.c2cpg.astcreation
 
 import io.joern.c2cpg.Config
 import io.joern.c2cpg.parser.HeaderFileFinder
+import io.joern.x2cpg.Ast
+import io.joern.x2cpg.AstCreatorBase
+import io.joern.x2cpg.ValidationMode
 import io.joern.x2cpg.datastructures.Stack.*
-import io.joern.x2cpg.{Ast, AstCreatorBase, ValidationMode, AstNodeBuilder as X2CpgAstNodeBuilder}
 import io.shiftleft.codepropertygraph.generated.NodeTypes
 import io.shiftleft.codepropertygraph.generated.nodes.*
-import io.shiftleft.semanticcpg.language.types.structure.NamespaceTraversal
-import org.eclipse.cdt.core.dom.ast.{IASTNode, IASTTranslationUnit}
-import org.slf4j.{Logger, LoggerFactory}
 import io.shiftleft.codepropertygraph.generated.DiffGraphBuilder
+import io.shiftleft.semanticcpg.language.types.structure.NamespaceTraversal
+import org.eclipse.cdt.core.dom.ast.IASTNode
+import org.eclipse.cdt.core.dom.ast.IASTTranslationUnit
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 
 import java.util.concurrent.ConcurrentHashMap
 import scala.collection.mutable
@@ -33,6 +37,7 @@ class AstCreator(
     with AstNodeBuilder
     with AstCreatorHelper
     with FullNameProvider
+    with TypeNameProvider
     with MacroHandler {
 
   protected implicit val schemaValidation: ValidationMode = config.schemaValidation
@@ -40,8 +45,6 @@ class AstCreator(
   protected val logger: Logger = LoggerFactory.getLogger(classOf[AstCreator])
 
   protected val scope: C2CpgScope = new C2CpgScope()
-
-  protected val usingDeclarationMappings: mutable.Map[String, String] = mutable.HashMap.empty
 
   // TypeDecls with their bindings (with their refs) for lambdas and methods are not put in the AST
   // where the respective nodes are defined. Instead, we put them under the parent TYPE_DECL in which they are defined.
