+++ joern-cli/frontends/c2cpg/src/main/scala/io/joern/c2cpg/astcreation/AstForFunctionsCreator.scala
@@ -315,7 +315,7 @@ trait AstForFunctionsCreator { this: AstCreator =>
         (
           s.getDeclarators.headOption
             .map(n => ASTStringUtil.getSimpleName(n.getName))
-            .getOrElse(uniqueName("", "", "param")._1),
+            .getOrElse(fileLocalUniqueName("", "", "param")._1),
           code(s),
           cleanType(typeForDeclSpecifier(s)),
           false
